.docx file as an attcahment fix
