+++ apps/sim/app/api/tools/gmail/send/route.ts
@@ -6,7 +6,7 @@ import { generateRequestId } from '@/lib/core/utils/request'
 import { withRouteHandler } from '@/lib/core/utils/with-route-handler'
 import { RawFileInputArraySchema } from '@/lib/uploads/utils/file-schemas'
 import { processFilesToUserFiles } from '@/lib/uploads/utils/file-utils'
-import { downloadFileFromStorage } from '@/lib/uploads/utils/file-utils.server'
+import { downloadFileForDelivery } from '@/lib/uploads/utils/file-utils.server'
 import { renderAgentResponseToString } from '@/tools/gmail/markUpRenderUtil'
 import {
   base64UrlEncode,
@@ -137,6 +137,8 @@ export const POST = withRouteHandler(async (request: NextRequest) => {
           )
         }
 
+        const ownerKey = authResult.userId ? `user:${authResult.userId}` : undefined
+
         const attachmentBuffers = await Promise.all(
           attachments.map(async (file) => {
             try {
@@ -144,11 +146,16 @@ export const POST = withRouteHandler(async (request: NextRequest) => {
                 `[${requestId}] Downloading attachment: ${file.name} (${file.size} bytes)`
               )
 
-              const buffer = await downloadFileFromStorage(file, requestId, logger)
+              const { buffer, contentType } = await downloadFileForDelivery(
+                file,
+                requestId,
+                logger,
+                { ownerKey, signal: request.signal }
+              )
 
               return {
                 filename: file.name,
-                mimeType: file.type || 'application/octet-stream',
+                mimeType: contentType || file.type || 'application/octet-stream',
                 content: buffer,
               }
             } catch (error) {
